* fixed freebsd
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4352 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-shared/ByteBuffer.h
@@ -377,7 +377,7 @@ public:
 	void hexlike()
 		{
 			uint32 j = 1, k = 1;
-			printf("STORAGE_SIZE: %u\n", size() );
+			printf("STORAGE_SIZE: %u\n", (unsigned int)size() );
 			for(uint32 i = 0; i < size(); i++)
 			{
 				if ((i == (j*8)) && ((i != (k*16))))
